revert changes of app + use access gate
